requirements.txt
flask
line-bot-sdk
openai
pyngrok
